requirements.txt
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client==2.169.0
google-auth==2.40.1
pytest==8.3.5
psycopg[binary]==3.2.9
